test: merge topology module tests
